File Rearrangement
Moved, removed and configured files adjusting for a different file setup. Code initializes and runs as expected
