+++ App/Backend/config.py
@@ -1,10 +1,10 @@
 import os
 
 def load_config(app, overrides):
-    if os.path.exists(os.path.join('./App', 'custom_config.py')):
-        app.config.from_object('App.custom_config')
+    if os.path.exists(os.path.join('./App/Backend', 'custom_config.py')):
+        app.config.from_object('App.Backend.custom_config')
     else:
-        app.config.from_object('App.default_config')
+        app.config.from_object('App.Backend.default_config')
     app.config.from_prefixed_env()
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
     app.config['TEMPLATES_AUTO_RELOAD'] = True
