+++ App/Backend/controllers/initialize.py
@@ -1,5 +1,5 @@
 from .user import create_user
-from App.database import db
+from App.Backend.database import db
 
 
 def initialize():
