+++ App/Backend/models/user.py
@@ -1,5 +1,5 @@
 from werkzeug.security import check_password_hash, generate_password_hash
-from App.database import db
+from App.Backend.database import db
 
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
