+++ App/Backend/tests/test_app.py
@@ -1,10 +1,10 @@
 import os, tempfile, pytest, logging, unittest
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from App.main import create_app
-from App.database import db, create_db
-from App.models import User
-from App.controllers import (
+from App.Backend.main import create_app
+from App.Backend.database import db, create_db
+from App.Backend.models import User
+from App.Backend.controllers import (
     create_user,
     get_all_users_json,
     login,
