+++ App/Backend/views/user.py
@@ -1,0 +1,25 @@
+from flask import Blueprint, render_template, jsonify, request, send_from_directory, flash, redirect, url_for
+from flask_jwt_extended import jwt_required, current_user as jwt_current_user
+
+from.index import index_views
+
+from App.Backend.controllers import (
+    create_user,
+    get_all_users,
+    get_all_users_json,
+    jwt_required
+)
+
+user_views = Blueprint('user_views', __name__)
+
+@user_views.route('/users', methods=['GET'])
+def get_user_page():
+    users = get_all_users()
+    return
+
+@user_views.route('/users', methods=['POST'])
+def create_user_action():
+    data = request.form
+    flash(f"User {data['username']} created!")
+    create_user(data['username'], data['password'])
+    return
