--- App/templates/admin/index.html
@@ -1,5 +1,0 @@
-{% extends 'admin/master.html' %}
-
-{% block body %}
-  <p>Hello world</p>
-{% endblock %}
