--- App/templates/message.html
@@ -1,15 +1,0 @@
-{% extends "layout.html" %}
-{% block title %}{{title}}{% endblock %}
-{% block page %}{{title}}{% endblock %}
-
-{{ super() }}
-
-{% block content %}
-<div class="container">
-    <div class="row" style="margin-top: 10%;">
-        <div class="col s12 m10 offset-m1">
-            <p>{{message}}</p>
-        </div>
-    </div>
-</div>
-{% endblock %}
