+++ wsgi.py
@@ -2,10 +2,10 @@ import click, pytest, sys
 from flask import Flask
 from flask.cli import with_appcontext, AppGroup
 
-from App.database import db, get_migrate
-from App.models import User
-from App.main import create_app
-from App.controllers import ( create_user, get_all_users_json, get_all_users, initialize )
+from App.Backend.database import db, get_migrate
+from App.Backend.models import User
+from App.Backend.main import create_app
+from App.Backend.controllers import ( create_user, get_all_users_json, get_all_users, initialize )
 
 app = create_app()
 migrate = get_migrate(app)
